### 2018-12-05 #1117 @whedon
[PRE REVIEW]: Cadabra2: computer algebra for field theory revisited
**Submitting author:** @kpeeters (<a href="http://orcid.org/0000-0002-3077-8193">Kasper Peeters</a>)
**Repository:** <a href="https://github.com/kpeeters/cadabra2" target ="_blank">https://github.com/kpeeters/cadabra2</a>
**Version:** 2.2.2
