### 2018-12-05 #1117 @danielskatz
@kpeeters - you should put the paper contents (md, bib) in a paper directory, then add a new comment here with the line
```@whedon generate pdf```
to build it
