### 2018-12-05 #1117 @danielskatz
It looks like there might be some reference issue that you can clean up, but we should also start getting reviewers.  Do you have any suggestions for potential reviewers? [This list of people](https://bit.ly/joss-reviewers) have already agreed to review for JOSS and may be suitable.
