### 2018-12-05 a6896b5
Paper fixes
