+++ paper/paper.bib
@@ -1,6 +1,7 @@
 @article{cadabra_hep,
   author  = "Peeters, Kasper",
   title   = "{Introducing Cadabra: a symbolic computer algebra system for field theory problems}",
+  year    = "2007",
   eprint  = "hep-th/0701238"
 }
 @article{cadabra_cpc,
@@ -24,7 +25,7 @@
   author = {Wenzel Jakob and Jason Rhinelander and Dean Moldovan},
   year = {2017},
   note = {https://github.com/pybind/pybind11},
-  title = {pybind11 -- Seamless operability between C++11 and Python}
+  title = {pybind11 -- Seamless operability between {C++11} and {Python}}
 }
 @article{butter,
   title={The component structure of conformal supergravity invariants in six dimensions},
@@ -57,7 +58,7 @@
   publisher={Elsevier}
 }
 @article{durkee,
-  title={Generalization of the Geroch--Held--Penrose formalism to higher dimensions},
+  title={Generalization of the {Geroch-Held-Penrose} formalism to higher dimensions},
   author={Durkee, Mark and Pravda, Vojt{\v{e}}ch and Pravdov{\'a}, Alena and Reall, Harvey S},
   journal={Classical and Quantum Gravity},
   volume={27},
@@ -67,7 +68,7 @@
   publisher={IOP Publishing}
 }
 @article{koile,
-  title={Hadron structure functions at small x from string theory},
+  title={Hadron structure functions at small {$x$} from string theory},
   author={Koile, Ezequiel and Kovensky, Nicolas and Schvellinger, Martin},
   journal={Journal of High Energy Physics},
   volume={2015},
@@ -77,7 +78,7 @@
   publisher={Springer}
 }
 @article{geissbuhler,
-  title={Double field theory and $$\backslash$mathcal $\{$N$\}$=$\{$4$\}$ $ gauged supergravity},
+  title={Double field theory and N=4 gauged supergravity},
   author={Geissb{\"u}hler, David},
   journal={Journal of High Energy Physics},
   volume={2011},
@@ -87,7 +88,7 @@
   publisher={Springer}
 }
 @article{christensen,
-  title={Boundary stress-energy tensor and Newton-Cartan geometry in Lifshitz holography},
+  title={Boundary stress-energy tensor and {Newton-Cartan} geometry in {Lifshitz} holography},
   author={Christensen, Morten H and Hartong, Jelle and Obers, Niels A and Rollier, Blaise},
   journal={Journal of High Energy Physics},
   volume={2014},
@@ -97,7 +98,7 @@
   publisher={Springer}
 }
 @article{sympy,
-  title = {SymPy: symbolic computing in Python},
+  title = {{SymPy}: symbolic computing in {Python}},
   author = {Meurer, Aaron et al.},
   year = 2017,
   month = jan,
@@ -109,7 +110,8 @@
   doi = {10.7717/peerj-cs.103}
 }
 @misc{mathematica,
-  author = {Wolfram Research{,} Inc.},
+  author = {Wolfram Research Inc.},
   title = {Mathematica},
-  note = {Champaign, IL, 2018}
+  year = {2018},
+  note = {Champaign, IL}
 }
