+++ paper/paper.bib
@@ -110,7 +110,7 @@
   doi = {10.7717/peerj-cs.103}
 }
 @misc{mathematica,
-  author = {Wolfram Research Inc.},
+  author = {Wolfram~Research~Inc.},
   title = {Mathematica},
   year = {2018},
   note = {Champaign, IL}
