### 2018-12-05 #1117 @kpeeters
Am fixing the refs (is there a recommended tool to generate pdf locally?).

Will have a scan through that list, will get back to you asap.
