+++ paper/paper.bib
@@ -110,8 +110,8 @@
   doi = {10.7717/peerj-cs.103}
 }
 @misc{mathematica,
-  author = {Wolfram~Research~Inc.},
-  title = {Mathematica},
-  year = {2018},
-  note = {Champaign, IL}
+  author = "{Wolfram Research Inc.}",
+  title = "Mathematica",
+  year = "2018",
+  note = "Champaign, IL"
 }
