### 2018-12-05 #1117 @danielskatz
https://github.com/openjournals/joss/issues/379 might be helpful for understanding the paper generation process

👋 @arfon - maybe we need to move some of this into our docs, rather than leaving it as an issue?
