### 2018-12-05 c04fc38
Fix bibliography
