+++ paper/paper.bib
@@ -25,7 +25,7 @@
   author = {Wenzel Jakob and Jason Rhinelander and Dean Moldovan},
   year = {2017},
   note = {https://github.com/pybind/pybind11},
-  title = {pybind11 -- Seamless operability between {C++11} and {Python}}
+  title = {{pybind11} -- {Seamless} operability between {C++11} and {Python}}
 }
 @article{butter,
   title={The component structure of conformal supergravity invariants in six dimensions},
@@ -78,7 +78,7 @@
   publisher={Springer}
 }
 @article{geissbuhler,
-  title={Double field theory and N=4 gauged supergravity},
+  title={Double field theory and {$N=4$} gauged supergravity},
   author={Geissb{\"u}hler, David},
   journal={Journal of High Energy Physics},
   volume={2011},
