### 2018-12-05 #1117 @kpeeters
@arfon @danielskatz Definitely put that in your docs please.
