### 2018-12-05 #1117 @kpeeters
As for reviewer suggestions: andrewfowlie, StewMH, rljacobson, moorepants, henrik227.
