### 2018-12-05 #1117 @danielskatz
👋 @andrewfowlie, @StewMH, @moorepants, @henrik227

Would any of you be willing to review this JOSS submission?
