### 2018-12-05 #1117 @danielskatz
@kpeeters - do all the references that have DOIs lists them in the bib?
