### 2018-12-05 #1117 @StewMH
Hi @danielskatz,

I'd be happy to review this - my first time as a reviewer.

Cheers,
Stewart
