### 2018-12-05 #1117 @kpeeters
@danielskatz Will check all entries to see if any DOIs can be added (not a very common way to cite in my field).
