### 2018-12-06 #1117 @henrik227
Hi @danielskatz,

I can review this too.

Henrik
