### 2018-12-06 #1117 @danielskatz
Thanks @StewMH and @henrik227 - I'll add you and start the review
