### 2018-12-10 5a32e3f
Bibliography updates
