+++ paper/paper.bib
@@ -1,117 +1,130 @@
 @article{cadabra_hep,
-  author  = "Peeters, Kasper",
-  title   = "{Introducing Cadabra: a symbolic computer algebra system for field theory problems}",
-  year    = "2007",
-  eprint  = "hep-th/0701238"
+  author    = "Peeters, Kasper",
+  title     = "{Introducing Cadabra: a symbolic computer algebra system for field theory problems}",
+  year      = "2007",
+  eprint    = "hep-th/0701238"
 }
 @article{cadabra_cpc,
   author    = "Peeters, Kasper",
   title     = "A field-theory motivated approach to symbolic computer algebra",
   year      = "2006",
-  eprint    = "cs.sc/0608005",
   journal   = "Comp.Phys.Commun.",
   volume    = "176",
-  pages     = "550-558"
+  pages     = "550-558",
+  eprint    = "cs.sc/0608005",
+  doi       = "10.1016/j.cpc.2007.01.003"
 }
 @article{cadabra_zenodo,
-  title={Cadabra: a field-theory motivated approach to computer algebra},
-  DOI={10.5281/zenodo.1842262},
-  publisher={Zenodo},
-  author={Kasper Peeters},
-  year={2018},
-  month={Dec}
+  title     = "Cadabra: a field-theory motivated approach to computer algebra",
+  publisher = "Zenodo",
+  author    = "Kasper Peeters",
+  year      = "2018",
+  doi       = "10.5281/zenodo.1842262"
 }
 @misc{pybind11,
-  author = {Wenzel Jakob and Jason Rhinelander and Dean Moldovan},
-  year = {2017},
-  note = {https://github.com/pybind/pybind11},
-  title = {{pybind11} -- {Seamless} operability between {C++11} and {Python}}
+  author    = "Wenzel Jakob and Jason Rhinelander and Dean Moldovan",
+  year      = "2017",
+  note      = "https://github.com/pybind/pybind11",
+  title     = "{pybind11} -- {Seamless} operability between {C++11} and {Python}"
 }
 @article{butter,
-  title={The component structure of conformal supergravity invariants in six dimensions},
-  author={Butter, Daniel and Novak, Joseph and Tartaglino-Mazzucchelli, Gabriele},
-  journal={Journal of High Energy Physics},
-  volume={2017},
-  number={5},
-  pages={133},
-  year={2017},
-  publisher={Springer}
+  title     = {The component structure of conformal supergravity invariants in six dimensions},
+  author    = {Butter, Daniel and Novak, Joseph and Tartaglino-Mazzucchelli, Gabriele},
+  journal   = {Journal of High Energy Physics},
+  volume    = {2017},
+  number    = {5},
+  pages     = {133},
+  year      = {2017},
+  publisher = {Springer},
+  eprint    = {},
+  doi       = {10.1007/JHEP05(2017)133}
 }
 @article{malik,
-  title={Cosmological perturbations},
-  author={Malik, Karim A and Wands, David},
-  journal={Physics Reports},
-  volume={475},
-  number={1-4},
-  pages={1--51},
-  year={2009},
-  publisher={Elsevier}
+  title     = {Cosmological perturbations},
+  author    = {Malik, Karim A and Wands, David},
+  journal   = {Physics Reports},
+  volume    = {475},
+  number    = {1-4},
+  pages     = {1--51},
+  year      = {2009},
+  publisher = {Elsevier},
+  eprint    = {},
+  doi       = {10.1016/j.physrep.2009.03.001}
 }
 @article{buchel,
-  title={Universal holographic hydrodynamics at finite coupling},
[…]
