### 2018-12-10 27b4e95
Add DOIs to all paper references
