+++ paper/paper.bib
@@ -36,7 +36,7 @@
   pages     = {133},
   year      = {2017},
   publisher = {Springer},
-  eprint    = {},
+  eprint    = {arXiv:1701.08163},
   doi       = {10.1007/JHEP05(2017)133}
 }
 @article{malik,
@@ -48,7 +48,7 @@
   pages     = {1--51},
   year      = {2009},
   publisher = {Elsevier},
-  eprint    = {},
+  eprint    = {arXiv:0809.4944},
   doi       = {10.1016/j.physrep.2009.03.001}
 }
 @article{buchel,
@@ -72,8 +72,8 @@
   pages     = {215010},
   year      = {2010},
   publisher = {IOP Publishing},
-  eprint    = {},
-  doi       = {}
+  eprint    = {arXiv:1002.4826},
+  doi       = {10.1088/0264-9381/27/21/215010}
 }
 @article{koile,
   title     = {Hadron structure functions at small {$x$} from string theory},
@@ -84,8 +84,8 @@
   pages     = {1},
   year      = {2015},
   publisher = {Springer},
-  eprint    = {},
-  doi       = {}
+  eprint    = {arXiv:1412.6509},
+  doi       = {10.1007/JHEP05(2015)001}
 }
 @article{geissbuhler,
   title     = {Double field theory and {$N=4$} gauged supergravity},
@@ -96,8 +96,8 @@
   pages     = {116},
   year      = {2011},
   publisher = {Springer},
-  eprint    = {},
-  doi       = {}
+  eprint    = {arXiv:1109.4280},
+  doi       = {10.1007/JHEP11(2011)116}
 }
 @article{christensen,
   title     = {Boundary stress-energy tensor and {Newton-Cartan} geometry in {Lifshitz} holography},
@@ -108,8 +108,8 @@
   pages     = {57},
   year      = {2014},
   publisher = {Springer},
-  eprint    = {},
-  doi       = {}
+  eprint    = {arXiv:1311.6471},
+  doi       = {10.1007/JHEP01(2014)057}
 }
 @article{sympy,
   title     = {{SymPy}: symbolic computing in {Python}},
