### 2018-12-10 #1118 @danielskatz
👋 @StewMH @henrik227 - I hope you can get started soon, at least by checking the CoI and CoC boxes...
