### 2018-12-10 6b5806c
Make ImplicitIndex much more flexible.
