+++ core/properties/ImplicitIndex.cc
@@ -15,10 +15,10 @@ bool ImplicitIndex::parse(Kernel&, keyval_t& keyvals)
 	while(ki!=keyvals.end()) {
 //		std::cout << "ImplicitIndex: " << ki->first << " = " << *ki->second->name << std::endl;
 		if(ki->first=="name") {
-			if(*ki->second->multiplier!=1) {
-				throw std::logic_error("ImplicitIndex: use quotes to label names when they start with a number.");
-				}
-			set_names.push_back(*ki->second->name);
+			throw std::logic_error("ImplicitIndex: argument 'name' no longer supported");
+			}
+		else if(ki->first=="explicit") {
+			explicit_form=ki->second;
 			}
 		else throw ConsistencyException("Property 'ImplicitIndex' does not accept key '"+ki->first+"'.");
 		++ki;
@@ -30,8 +30,8 @@ bool ImplicitIndex::parse(Kernel&, keyval_t& keyvals)
 void ImplicitIndex::latex(std::ostream& str) const
 	{
 	property::latex(str);
-	for(size_t n=0; n<set_names.size(); ++n) {
-		if(n>0) str << ", ";
-		str << set_names[n];
-		}
+//	for(size_t n=0; n<set_names.size(); ++n) {
+//		if(n>0) str << ", ";
+//		str << set_names[n];
+//		}
 	}
