### 2018-12-13 #1118 @henrik227
The only thing I can find regarding guidelines for contributors is [this](https://cadabra.science/developers.html). @danielskatz Is that considered sufficient?
