### 2018-12-13 f619659
Cleanup doxygen documentation. Add untrace.
