### 2018-12-13 8a41782
Fix for untrace of traces with single-factor argument.
