### 2018-12-16 #1118 @danielskatz
@henrik227 - it's not ideal, but it almost meets the minimum required for JOSS.  The one thing that I think is needed is to have this in the README as well.  If you have more suggestions, please make them as an issue in the source repository, but they will be treated as suggestions, not a requirement for JOSS acceptance.
