### 2018-12-17 #1118 @kpeeters
Am of course happy to expand this text and add to the README; this is trivial compared to writing the software and getting the documentation/tutorials/packages organised...
