### 2018-12-17 #1118 @StewMH
@danielskatz I went through the review process today - everything looks good except for a minor issue in the unit tests:
https://github.com/kpeeters/cadabra2/issues/122

Cheers,

Stewart
