### 2018-12-17 326dd82
Fix issue with single-term recognition
