### 2018-12-17 kpeeters/cadabra2#122 @kpeeters
I have pushed a fix (tested on 16.04), can you please try again?
