### 2018-12-19 #1118 @danielskatz
@kpeeters - please let us know when you think the community guidelines section is complete, so @henrik227 can re-review this and check off his final item.

The other needed item is to resolve https://github.com/kpeeters/cadabra2/issues/122
