### 2018-12-19 af90b7d
Extend 'how to contribute' guidelines.
