### 2018-12-19 #1118 @kpeeters
@danielskatz - please also let me know if you want me to tag the current state of the repo with a new version.
