### 2018-12-19 #1118 @danielskatz
Thanks @kpeeters

@henrik227 - please see what you think.
@StewMH - do you think the unit test issue is something that should block acceptance?
