### 2018-12-19 #1118 @danielskatz
> @danielskatz - please also let me know if you want me to tag the current state of the repo with a new version.

Once we are done with the reviews, you can either update the current version or tag it as a new one.
