### 2018-12-19 #1118 @henrik227
> @henrik227 - please see what you think.

I'm fine with those guidelines.
