### 2018-12-19 #1118 @StewMH
Hi @danielskatz, @kpeeters,

I can reproduce the bug in a clean Docker-based setup: I will post the details to the main ticket. I don't think it's a blocker for acceptance but it seems to suggest a CI system should be set up.

Cheers,

Stewart
