### 2018-12-19 #1118 @kpeeters
Volunteers for CI integration welcome ;-)  I had things running on Travis at some point but it started eating too much into my time when they kept changing the system under my feet. I now just have a bunch of VMs that I use to build binary packages from time to time. The Python ecosystem is a pain to adapt to...
