### 2018-12-19 kpeeters/cadabra2#122 @StewMH
Hi,

I can reproduce this in an Ubuntu build under Docker, which seems to rule out issues with my environment:
```
docker run -it stewartmh/cadabra2-build:latest
cd cadabra2/build
ctest
```

This Docker container could form a decent basis for future CI - you can certainly run Travis with a Docker container but I don't know the details.

Cheers,

Stewart
