### 2018-12-21 #1118 @danielskatz
I think this is ready to accept

@kpeeters - can you archive the current version (and change the version number if needed, in the paper source as well) and let me know the DOI?
