### 2018-12-21 c0d10e0
Fix Derivative import from sympy
