### 2018-12-21 kpeeters/cadabra2#122 @kpeeters
This turned out to be caused by newer versions of sympy writing multiple derivatives as `Derivative(f(x), (x,2))` instead of `Derivative(f(x),x,x)`. I have pushed a fix for this which solves this.
