### 2018-12-22 #1118 @kpeeters
@danielskatz - I have archived the current version as 2.2.4, which has DOI 10.5281/zenodo.2500762.
