### 2018-12-22 #1118 @danielskatz
@StewMH, @henrik227 - thanks very much for your reviews!
