### 2018-12-22 #1118 @danielskatz
@kpeeters - your paper is now accepted into JOSS ⚡️🚀💥
